moves code to util from leetcode.js
